requirements.txt
Django==4.2.7
djangorestframework==3.14.0
openai==1.3.7
google-api-python-client==2.100.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
python-telegram-bot==20.6
python-dotenv==1.0.0
django-cors-headers==4.3.1

